backend/spectra-prediction/code/requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
Flask-RESTful==0.3.9
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.21.4
pandas==1.3.4
pika==1.2.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
threadpoolctl==3.0.0
urllib3==1.26.7
Werkzeug==2.0.2
